+++ examples/rails-6.1.5/app/models/user.rb
@@ -1,0 +1,3 @@
+class User < ApplicationRecord
+  has_many :posts, inverse_of: :user, dependent: :destroy
+end